api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
asyncpg==0.30.0
httptools>=0.6.3
uvloop>=0.14.0
PyJWT>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
